Implementation performed on the CourseProcessing class. The algorithm handles a circular correlation by treating all the courses involved as co-requisites.
